+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # Changelog
 
+## Unreleased
+ - Turn ASTs into queries via `lucene.toString(ast)`.
+ - Retain information about quoted strings in the AST.
+
 ## 1.0.0
  - Initial release